+++ source/extensions/omni.flux.validator.mass.core/omni/flux/validator/mass/core/tests/unit/fake_plugins.py
@@ -150,6 +150,10 @@ class FakeContext(_ContextBase):
         """
         ui.Label("Fake context label", alignment=ui.Alignment.CENTER)
 
+    @omni.usd.handle_exception
+    async def _on_crash(self, schema_data: Any, data: Any) -> None:
+        pass
+
 
 class FakeSelector(_SelectorBase):
     class Data(_SelectorBase.Data):
@@ -239,6 +243,10 @@ class FakeSelector(_SelectorBase):
         Build the UI for the plugin
         """
         ui.Label("Fake selector label", alignment=ui.Alignment.CENTER)
+
+    @omni.usd.handle_exception
+    async def _on_crash(self, schema_data: Any, data: Any) -> None:
+        pass
 
 
 class FakeCheck(_CheckBase):
@@ -345,6 +353,10 @@ class FakeCheck(_CheckBase):
         Build the UI for the plugin
         """
         ui.Label("Fake check label", alignment=ui.Alignment.CENTER)
+
+    @omni.usd.handle_exception
+    async def _on_crash(self, schema_data: Any, data: Any) -> None:
+        pass
 
 
 def register_fake_plugins():