+++ source/extensions/lightspeed.trex.texture_replacements.core.shared/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.1.1]
+### Fixed
+- Fixed hot-reload by allowing reuse of the validators
+
 ## [1.1.0]
 ### Added
 - Added ability to filter get assets on a specific layer