+++ source/extensions/omni.flux.service.factory/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.3.0]
+### Changed
+- Use generic factory instead of service-specific factory
+
 ## [1.2.1]
 ### Fixed
 - Fixed documentation