+++ source/extensions/omni.flux.validator.mass.core/docs/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.11.10]
+### Fixed
+- Fixed test plugins to implement all abstract methods
+
 ## [1.11.9]
 ### Changed
 - Update deps