+++ source/extensions/omni.flux.validator.mass.widget/docs/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.7.3]
+### Fixed
+- Fixed test plugins to implement all abstract methods
+
 ## [1.7.2] - 2024-04-29
 ### Fixed
 - Fixed ingestion queue scroll bar