+++ source/extensions/omni.flux.validator.plugin.selector.usd/docs/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.8.2]
+### Fixed
+- Fixed test plugins to implement all abstract methods
+
 ## [1.8.1]
 ### Fixed
 - Fix `AllTextures` plugin